requirements.txt
pandas==1.2.2
pyyaml==5.4.1
mypy==0.812
textblob-de==0.4.3
flake8==3.9.0
nose-py3==1.4.0
num2words==0.5.10